Feature/binary protocol (#75)
* Add common module. Add channel initializer.

* Add electra client interface to specify client API.

* Add pipeline utils.

* Add correct client dependencies.

* Add thread factory.

* Add factory class and basic client implementation.

* Add connect logic.

* Add consumers to the get methods.

* Add binary handler logic for GET action.

* Implement logic for the get methods.

* Add koloboke dependency.

* Add server side binary protocol implementation.

* Add compile plugin.

* Update pom.xml

* Update pom.xml

* Add length check.

* Delete duplicated class.

* Add put logic.

* Add remove logic.

* Do not write null bytes.

* Start to implement update.

* Complete binary protocol update implementation.

* Correct logging.

* Fix update.

* Use enum for the binary protocol action.

* Add channel inactive log.

* Add 'USE_STORAGE' action handling.

* Remove singleton. Finish log for 'USE_STORAGE' action. Add full multi database support for binary protocol.

* Add length checks for storage name.

+++ server/src/main/java/io/electra/server/binary/ElectraServerHandler.java
@@ -105,7 +105,11 @@ public class ElectraServerHandler extends SimpleChannelInboundHandler<ByteBuf> {
                     break;
                 case CREATE_STORAGE:
                     int createStorageNameLength = byteBuf.readInt();
-                    // TODO: 16.12.2017 Limit size
+                    if (createStorageNameLength > 128) {
+                        // TODO: 05.01.2018 Return error code?
+                        logger.error("Storage name to create cannot be longer than 128 bytes");
+                        break;
+                    }
                     byte[] createStorageNameBytes = new byte[createStorageNameLength];
 
                     // This will create the database and use it for further actions
@@ -113,7 +117,11 @@ public class ElectraServerHandler extends SimpleChannelInboundHandler<ByteBuf> {
                     break;
                 case DELETE_STORAGE:
                     int deleteStorageNameLength = byteBuf.readInt();
-                    // TODO: 16.12.2017 Limit size
+                    if (deleteStorageNameLength > 128) {
+                        // TODO: 05.01.2018 Return error code?
+                        logger.error("Storage name to delete cannot be longer than 128 bytes");
+                        break;
+                    }
                     byte[] deleteStorageNameBytes = new byte[deleteStorageNameLength];
 
                     electraCore.deleteDatabase(new String(deleteStorageNameBytes, Charsets.UTF_8));
